refactor: Add cinemaId parameter to listCombos API function
This commit modifies the listCombos API function in the listCombos.ts file to accept an optional cinemaId parameter. If the cinemaId is provided, it will be included as a query parameter in the API request to retrieve combos specific to the cinema. If no cinemaId is provided, an empty array will be returned. This change allows for more targeted retrieval of combos based on cinema, enhancing the flexibility and customization of the API function.