refactor: Add skeleton loading for cinema and date selection
This commit adds skeleton loading components to the CalendarManagement and PopcornManagement components. When the cinemaSelected state is falsy, the cinema and date selection components will display skeleton loading buttons to indicate that the data is being loaded. This improves the user experience by providing visual feedback and preventing user interaction until the data is ready.